requirements.txt
pandas
scikit-learn==1.4.1.post1
flask==3.0.2
requests==2.31.0
matplotlib==3.8.3
pyyaml==6.0.1
beautifulsoup4==4.12.3
matplotlib
numpy==1.26.2
scipy==1.13.1
pyyaml==6.0.1
flask-socketio==5.3.6